+++ app/uk/gov/hmrc/tai/model/domain/IabdSummary.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.model.domain
+
+import play.api.libs.json.{JsResult, JsSuccess, JsValue, Reads}
+
+case class IabdSummary(componentType: Int, employmentId: Option[Int], amount: BigDecimal)
+
+object IabdSummary {
+  val iabdSummaryReads = new Reads[IabdSummary] {
+    override def reads(json: JsValue): JsResult[IabdSummary] = {
+      val componentType = (json \ "type").as[Int]
+      val employmentId = (json \ "employmentId").asOpt[Int]
+      val amount = (json \ "amount").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      JsSuccess(IabdSummary(componentType, employmentId, amount))
+    }
+  }
+}
