+++ app/uk/gov/hmrc/tai/repositories/deprecated/TaxAccountSummaryRepository.scala
@@ -23,7 +23,7 @@ import uk.gov.hmrc.tai.connectors.TaxAccountConnector
 import uk.gov.hmrc.tai.model.domain.TaxOnOtherIncome.{taxAccountSummaryReads, taxOnOtherIncomeRead}
 import uk.gov.hmrc.tai.model.domain._
 import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent
-import uk.gov.hmrc.tai.model.domain.formatters.taxComponents.TaxAccountHodFormatters
+import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent.codingComponentReads
 import uk.gov.hmrc.tai.model.domain.taxAdjustments.TaxAdjustmentComponent.taxAdjustmentComponentReads
 import uk.gov.hmrc.tai.model.domain.taxAdjustments.{AlreadyTaxedAtSource, OtherTaxDue, ReliefsGivingBackTax, TaxAdjustment, TaxAdjustmentComponent, TaxReliefComponent}
 import uk.gov.hmrc.tai.model.tai.TaxYear
@@ -33,8 +33,7 @@ import scala.concurrent.{ExecutionContext, Future}
 @Singleton
 class TaxAccountSummaryRepository @Inject() (
   taxAccountConnector: TaxAccountConnector
-)(implicit ec: ExecutionContext)
-    extends TaxAccountHodFormatters {
+)(implicit ec: ExecutionContext) {
 
   def taxAccountSummary(nino: Nino, year: TaxYear)(implicit hc: HeaderCarrier): Future[BigDecimal] = {
     val componentTypesCanAffectTotalEst: Seq[TaxComponentType] =
